feature: discounted revenue for merchant invoice show page

+++ db/migrate/20230302232237_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.integer :quantity
-      t.integer :discount
+      t.float :discount
       t.references :merchant, foreign_key: true
 
       t.timestamps
